+++ mgt2e/module/helpers/templates.mjs
@@ -19,6 +19,8 @@ export const preloadHandlebarsTemplates = async function() {
         "systems/mgt2e/templates/actor/parts/actor-starship-combat.html",
         "systems/mgt2e/templates/actor/parts/actor-starship-hardware.html",
         "systems/mgt2e/templates/item/parts/item-general.html",
+        "systems/mgt2e/templates/item/parts/item-computer.html",
+        "systems/mgt2e/templates/item/parts/item-software.html",
         "systems/mgt2e/templates/item/parts/item-techLevel.html"
     ]);
 };